back_end/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
Pillow
websockets
aiohttp
asyncio